Move in memory ingest classes to ingest runner test jar
