+++ java/ingest/ingest-runner/src/test/java/sleeper/ingest/runner/testutils/InMemoryPartitionFileWriter.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package sleeper.systemtest.dsl.testutil.drivers;
+package sleeper.ingest.runner.testutils;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -29,7 +29,6 @@ import sleeper.core.statestore.FileReference;
 import sleeper.core.table.TableFilePaths;
 import sleeper.ingest.runner.impl.partitionfilewriter.PartitionFileWriter;
 import sleeper.ingest.runner.impl.partitionfilewriter.PartitionFileWriterFactory;
-import sleeper.ingest.runner.testutils.InMemorySketchesStore;
 import sleeper.sketches.Sketches;
 
 import java.util.ArrayList;
